Consolidate study session container into page
